+++ src/app/components/submission/short-answer-question.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { IQuestion } from 'src/app/models/iquestion.question';
+import { IQuestion } from 'src/app/models/iquestion-question';
 
 @Component({
   selector: 'short-answer-question',
